Use parso to parse imports
