+++ starkiller/pylsp_plugin/plugin.py
@@ -15,7 +15,7 @@ from pylsp import hookimpl  # type: ignore
 from pylsp.config.config import Config  # type: ignore
 from pylsp.workspace import Document, Workspace  # type: ignore
 
-from starkiller.parsing import ImportedName, find_from_import, find_import, parse_module
+from starkiller.parsing import ImportedName, ImportFromStatement, ImportModulesStatement, find_imports, parse_module
 from starkiller.project import StarkillerProject
 from starkiller.refactoring import get_attrs_as_names_edits, get_rename_edits
 
@@ -65,22 +65,34 @@ def pylsp_code_actions(
     aliases = plugin_settings.get("aliases", [])
 
     active_range = converter.structure(range, Range)
-    line = document.lines[active_range.start.line].rstrip("\r\n")
-    line_range = Range(
-        start=Position(line=active_range.start.line, character=0),
-        end=Position(line=active_range.start.line, character=len(line)),
+    line_no = active_range.start.line + 1
+
+    import_statement = find_imports(document.source, line_no)
+    if import_statement is None:
+        return []
+    import_range = Range(
+        start=Position(
+            line=import_statement.import_range.start.line - 1,
+            character=import_statement.import_range.start.char,
+        ),
+        end=Position(
+            line=import_statement.import_range.end.line - 1,
+            character=import_statement.import_range.end.char,
+        ),
     )
 
-    from_module, imported_names = find_from_import(line)
-    imported_modules = find_import(line)
-
-    if from_module and imported_names:
-        if any(name.name == "*" for name in imported_names):
-            code_actions.extend(get_ca_for_star_import(document, project, from_module, line_range, aliases))
+    if isinstance(import_statement, ImportFromStatement):
+        if import_statement.is_star:
+            code_actions.extend(
+                get_ca_for_star_import(document, project, import_statement.module, import_range, aliases)
+            )
         else:
-            code_actions.extend(get_ca_for_from_import(document, from_module, imported_names, line_range, aliases))
-    elif imported_modules:
-        code_actions.extend(get_ca_for_module_import(document, imported_modules, line_range))
+            imported_names = import_statement.names or set()
+            code_actions.extend(
+                get_ca_for_from_import(document, import_statement.module, imported_names, import_range, aliases)
+            )
+    elif isinstance(import_statement, ImportModulesStatement):
+        code_actions.extend(get_ca_for_module_import(document, import_statement.modules, import_range))
 
     return converter.unstructure(code_actions)
 
@@ -89,23 +101,23 @@ def get_ca_for_star_import(
     document: Document,
     project: StarkillerProject,
     from_module: str,
-    import_line_range: Range,
+    import_range: Range,
     aliases: dict,
 ) -> list[CodeAction]:
     undefined_names = parse_module(document.source, check_internal_scopes=True).undefined
     if not undefined_names:
-        return [get_ca_remove_unnecessary_import(document, import_line_range)]
+        return [get_ca_remove_unnecessary_import(document, import_range)]
 
     externaly_defined = project.find_definitions(from_module, set(undefined_names))
     if not externaly_defined:
-        return [get_ca_remove_unnecessary_import(document, import_line_range)]
+        return [get_ca_remove_unnecessary_import(document, import_range)]
 
-    text_edits_from = get_edits_replace_module_w_from(from_module, externaly_defined, import_line_range)
+    text_edits_from = get_edits_replace_module_w_from(from_module, externaly_defined, import_range)
     text_edits_module = get_edits_replace_from_w_module(
         document.source,
         from_module,
         {ImportedName(name) for name in externaly_defined},
-        import_line_range,
+        import_range,
         aliases,
     )
 
@@ -125,8 +137,8 @@ def get_ca_for_star_import(
 
 def get_ca_for_module_import(
     document: Document,
-    imported_modules: list[ImportedName],
-    line_range: Range,
+    imported_modules: set[ImportedName],
+    import_range: Range,
 ) -> list[CodeAction]:
     parsed = parse_module(document.source, check_internal_scopes=True, collect_imported_attrs=True)
 
@@ -135,16 +147,14 @@ def get_ca_for_module_import(
         # manually or with some other tool like Ruff
         return []
 
-    module = imported_modules[0]
+    module = imported_modules.pop()
     used_attrs = parsed.imported_attr_usages.get(module.alias or module.name)
     if not used_attrs:
-        return [get_ca_remove_unnecessary_import(document, line_range)]
+        return [get_ca_remove_unnecessary_import(document, import_range)]
 
-    text_edits = get_edits_replace_module_w_from(module.name, used_attrs, line_range)
+    text_edits = get_edits_replace_module_w_from(module.name, used_attrs, import_range)
 
-    for edit_range, new_value in get_attrs_as_names_edits(
-        document.source, module.alias or module.name, used_attrs
-    ):
+    for edit_range, new_value in get_attrs_as_names_edits(document.source, module.alias or module.name, used_attrs):
         rename_range = Range(
             start=Position(line=edit_range.start.line, character=edit_range.start.char),
             end=Position(line=edit_range.end.line, character=edit_range.end.char),
@@ -161,19 +171,9 @@ def get_ca_for_module_import(
 
 
 def get_ca_for_from_import(
-    document: Document,
-    from_module: str,
-    imported_names: set[ImportedName],
-    import_line_range: Range,
-    aliases: dict,
+    document: Document, from_module: str, imported_names: set[ImportedName], import_range: Range, aliases: dict
 ) -> list[CodeAction]:
-    text_edits = get_edits_replace_from_w_module(
-        document.source,
-        from_module,
-        imported_names,
-        import_line_range,
-        aliases,
-    )
+    text_edits = get_edits_replace_from_w_module(document.source, from_module, imported_names, import_range, aliases)
     return [
         CodeAction(
             title="Starkiller: Replace from import with module import",
@@ -183,28 +183,24 @@ def get_ca_for_from_import(
     ]
 
 
-def get_edits_replace_module_w_from(
-    from_module: str,
-    names: set[str],
-    import_line_range: Range,
-) -> list[TextEdit]:
+def get_edits_replace_module_w_from(from_module: str, names: set[str], import_range: Range) -> list[TextEdit]:
     names_str = ", ".join(names)
     new_text = f"from {from_module} import {names_str}"
-    return [TextEdit(range=import_line_range, new_text=new_text)]
+    return [TextEdit(range=import_range, new_text=new_text)]
 
 
 def get_edits_replace_from_w_module(
     source: str,
     from_module: str,
     names: set[ImportedName],
-    import_line_range: Range,
+    import_range: Range,
     aliases: dict[str, str],
 ) -> list[TextEdit]:
     new_text = f"import {from_module}"
     if from_module in aliases:
         alias = aliases[from_module]
         new_text += f" as {alias}"
-    text_edits = [TextEdit(range=import_line_range, new_text=new_text)]
+    text_edits = [TextEdit(range=import_range, new_text=new_text)]
 
     rename_map = {n.alias or n.name: f"{from_module}.{n.name}" for n in names}
     for edit_range, new_value in get_rename_edits(source, rename_map):
@@ -216,11 +212,8 @@ def get_edits_replace_from_w_module(
     return text_edits
 
 
-def get_ca_remove_unnecessary_import(
-    document: Document,
-    import_line_range: Range,
-) -> CodeAction:
-    import_line_num = import_line_range.start.line
+def get_ca_remove_unnecessary_import(document: Document, import_range: Range) -> CodeAction:
+    import_line_num = import_range.start.line
     import_line = document.lines[import_line_num]
 
     if import_line != len(document.lines) - 1:
